+++ l2-frontend/src/construct/ResearchDetail.vue
@@ -1,104 +1,138 @@
 <template>
-  <div>
+  <div class="margin-root">
     <h4 class="header">
       Редактирование анализа - {{ researchShortTitle }}
     </h4>
-    <div class="margin-root">
-      <div class="research-detail">
-        <div>
-          <div class="margin">
-            <label
-              for="title"
-              class="research-detail-label"
-            >Полное наименование</label>
-            <input
-              id="title"
-              v-model="research.title"
-              class="form-control"
-            >
-          </div>
-          <div class="margin">
-            <label
-              for="shortTitle"
-              class="research-detail-label"
-            >Краткое наименование</label>
-            <input
-              id="shortTitle"
-              v-model="research.shortTitle"
-              class="form-control"
-            >
-          </div>
+    <div class="research-detail">
+      <div>
+        <div class="margin">
+          <label
+            for="title"
+            class="research-detail-label"
+          >Полное наименование</label>
+          <input
+            id="title"
+            v-model="research.title"
+            class="form-control"
+            placeholder="Введите полное наименование"
+          >
         </div>
-        <div class="code">
-          <div class="margin code-item">
-            <label
-              for="code"
-              class="research-detail-label"
-            >Код НМУ</label>
-            <input
-              id="code"
-              v-model="research.code"
-              class="form-control"
-            >
-          </div>
-          <div class="margin code-item">
-            <label
-              for="ecpCode"
-              class="research-detail-label"
-            >Код ЕЦП</label>
-            <input
-              id="ecpCode"
-              v-model="research.ecpCode"
-              class="form-control"
-            >
-          </div>
-          <div class="margin code-item">
-            <label
-              for="internalCode"
-              class="research-detail-label"
-            >Код внутренний</label>
-            <input
-              id="internalCode"
-              v-model="research.internalCode"
-              class="form-control"
-            >
-          </div>
-        </div>
-        <div>
-          <div class="margin">
-            <label
-              for="preparation"
-              class="research-detail-label"
-            >Подготовка</label>
-            <textarea
-              id="preparation"
-              class="form-control"
-              style="height: 90px"
-              rows="4"
-            />
-          </div>
+        <div class="margin">
+          <label
+            for="shortTitle"
+            class="research-detail-label"
+          >Краткое наименование</label>
+          <input
+            id="shortTitle"
+            v-model="research.shortTitle"
+            class="form-control"
+            placeholder="Введите краткое наименование"
+          >
         </div>
       </div>
-      <div class="research-fractions">
-        <h4 class="header">
-          <table class="table">
-            <colgroup>
-              <col width="30">
-              <col width="30">
-              <col>
-              <col width="30">
-              <col width="30">
-            </colgroup>
-            <tr
-              v-for="(research) in props.tube.researches"
-              :key="research.pk"
-            >
-              <td class="border">
-                ура
-              </td>
+      <div class="code">
+        <div class="margin code-item">
+          <label
+            for="code"
+            class="research-detail-label"
+          >Код НМУ</label>
+          <input
+            id="code"
+            v-model="research.code"
+            class="form-control"
+            placeholder="Введите код НМУ"
+          >
+        </div>
+        <div class="margin code-item">
+          <label
+            for="ecpCode"
+            class="research-detail-label"
+          >Код ЕЦП</label>
+          <input
+            id="ecpCode"
+            v-model="research.ecpCode"
+            class="form-control"
+            placeholder="Введите код ЕЦП"
+          >
+        </div>
+        <div class="margin code-item">
+          <label
+            for="internalCode"
+            class="research-detail-label"
+          >Код внутренний</label>
+          <input
+            id="internalCode"
+            v-model="research.internalCode"
+            class="form-control"
+            placeholder="Введите код Внутренний"
+          >
+        </div>
+      </div>
+      <div>
+        <div class="margin">
+          <label
+            for="preparation"
+            class="research-detail-label"
+          >Подготовка</label>
+          <textarea
+            id="preparation"
+            class="form-control"
+            style="height: 90px"
+            rows="4"
+            placeholder="Введите подгтовка (напр. 'Не требуется')"
+          />
+        </div>
+      </div>
+    </div>
+    <h4 class="header">
+      Фракции
+    </h4>
+    <div class="research-fractions">
+      <div
+        v-for="tube in research.tubes"
+        :key="tube.pk"
+        class="tube-group"
+      >
+        <Tube :tube="tube" />
+        <table class="table">
+          <colgroup>
+            <col>
+            <col width="300">
+            <col width="300">
+          </colgroup>
+          <thead>
+            <tr>
+              <th><strong>Фракция</strong></th>
+              <th><strong>Ед. измерения</strong></th>
+              <th><strong>Вар. комментариев</strong></th>
             </tr>
-          </table>
-        </h4>
+          </thead>
+          <tr
+            v-for="fraction in tube.fractions"
+            :key="fraction.pk"
+          >
+            <td class="padding-td">
+              <input
+                v-model="fraction.title"
+                class="form-control"
+                placeholder="Введите название фракции"
+              >
+            </td>
+            <td class="padding-td">
+              <input
+                v-model="fraction.unit"
+                class="form-control"
+                placeholder="Введите ед. изм"
+              >
+            </td>
+            <td class="padding-td">
+              <input
+                v-model="fraction.variants"
+                class="form-control"
+              >
+            </td>
+          </tr>
+        </table>
       </div>
       <div class="margin-bottom flex-right">
         <button
@@ -120,6 +154,7 @@ import {
 import { useStore } from '@/store';
 import * as actions from '@/store/action-types';
 import api from '@/api';
+import Tube from '@/construct/Tube.vue';
 
 const store = useStore();
 
@@ -133,10 +168,18 @@ const props = defineProps({
 });
 
 interface fractionsData {
+  pk: number,
   title: string,
   unit: string,
   variants: string[] | null,
   sortWeight: number,
+}
+
+interface tubeData {
+  pk: number,
+  title: string,
+  color: string,
+  fractions: fractionsData[],
 }
 
 interface researchData {
@@ -147,7 +190,7 @@ interface researchData {
   internalCode: number | null,
   ecpCode: number | null,
   preparation: string | null,
-  fractions: fractionsData[],
+  tubes: tubeData[]
 }
 const researchShortTitle = ref('');
 
@@ -160,12 +203,20 @@ const research = ref<researchData>({
   internalCode: -1,
   ecpCode: -1,
   preparation: '',
-  fractions: [
+  tubes: [
     {
+      pk: -1,
       title: '',
-      unit: '',
-      variants: null,
-      sortWeight: -1,
+      color: '',
+      fractions: [
+        {
+          pk: -1,
+          title: '',
+          unit: '',
+          variants: null,
+          sortWeight: -1,
+        },
+      ],
     }],
 });
 
@@ -202,8 +253,11 @@ onMounted(() => {
 <style scoped lang="scss">
 .research-detail {
   display: grid;
+  background-color: #fff;
+  border-radius: 4px;
   grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
-  margin-bottom: 15px;
+  padding: 10px 0;
+  box-shadow: 0 1px 3px rgb(0 0 0 / 12%), 0 1px 2px rgb(0 0 0 / 24%);
 }
 .code {
   display: flex;
@@ -211,7 +265,7 @@ onMounted(() => {
   align-content: flex-start;
 }
 .margin-root {
-  margin: 0 20px 0 20px;
+  margin: 0 10px 0 10px;
 }
 .margin {
   margin: 0 5px;
@@ -228,10 +282,27 @@ onMounted(() => {
   flex-basis: 145px;
 }
 .header {
-  margin: 10px 0 10px 37px;
+  margin: 10px 0 10px 17px;
 }
 .flex-right {
   display: flex;
   justify-content: end;
+}
+.tube-group {
+  margin-bottom: 10px;
+  background-color: #fff;
+  padding: 10px 5px;
+  border-radius: 4px;
+  box-shadow: 0 1px 3px rgb(0 0 0 / 12%), 0 1px 2px rgb(0 0 0 / 24%);
+}
+.table {
+  table-layout: fixed;
+  margin-bottom: 0;
+}
+.padding-td {
+  padding: 5px;
+}
+.border {
+  border: 1px solid #bbb;;
 }
 </style>
