+++ l2-frontend/src/construct/Tube.vue
@@ -1,137 +1,40 @@
 <template>
   <div
-    class="research"
+    class="tube-item"
   >
-    {{ 'Анализы' }}
-    <table class="table">
-      <colgroup>
-        <col width="30">
-        <col width="30">
-        <col>
-        <col width="30">
-        <col width="30">
-      </colgroup>
-      <tr
-        v-for="(research, idx) in props.tube.researches"
-        :key="research.pk"
-      >
-        <td class="border">
-          <div class="button">
-            <button
-              :class="isFirstRow(research.order) ? 'transparent-button-disabled' : 'transparent-button'"
-              :disabled="isFirstRow(research.order)"
-              @click="updateOrder(idx, research.pk, research.order,'dec_order')"
-            >
-              <i class="glyphicon glyphicon-arrow-up" />
-            </button>
-          </div>
-        </td>
-        <td class="border">
-          <div class="button">
-            <button
-              :class="isLastRow(research.order) ? 'transparent-button-disabled' : 'transparent-button'"
-              :disabled="isLastRow(research.order)"
-              @click="updateOrder(idx, research.pk, research.order, 'inc_order')"
-            >
-              <i class="glyphicon glyphicon-arrow-down" />
-            </button>
-          </div>
-        </td>
-        <td class="border research-title">
-          {{ research.title }}
-        </td>
-        <td class="border">
-          <div class="button">
-            <button
-              class="transparent-button"
-              @click="changeVisibility(research.pk)"
-            >
-              {{ research.hide ? 'пок' : 'скр' }}
-            </button>
-          </div>
-        </td>
-        <td class="border">
-          <div class="button">
-            <button
-              class="transparent-button"
-              @click="edit(research.pk)"
-            >
-              <i class="fa fa-pencil" />
-            </button>
-          </div>
-        </td>
-      </tr>
-    </table>
-    <div> {{ props.tube.tubes.length > 0 ? 'Ёмкости' : 'Ёмкости не привязаны' }}</div>
     <div
-      v-for="currentTube in props.tube.tubes"
-      :key="currentTube.pk"
-      class="tube-item"
+      class="sq"
     >
       <div
-        class="sq"
-      >
-        <div
-          class="color-sq"
-          :style="'background-color: ' + currentTube.color"
-        />
-      </div>
-      <div> {{ currentTube.title }}</div>
+        class="color-sq"
+        :style="'background-color: ' + props.tube.color"
+      />
     </div>
+    <div> {{ props.tube.title }}</div>
   </div>
 </template>
 
 <script setup lang="ts">
-import { computed, getCurrentInstance } from 'vue';
 
+import { PropType } from 'vue';
+
+interface tube {
+  pk: number,
+  title: string,
+  color: string,
+}
 const props = defineProps({
   tube: {
-    type: Object,
+    type: Object as PropType<tube>,
     required: true,
   },
 });
-const root = getCurrentInstance().proxy.$root;
-const emit = defineEmits(['updateOrder', 'changeVisibility', 'edit']);
-
-const minMaxOrder = computed(() => {
-  let min = 0;
-  let max = 0;
-  for (const row of props.tube.researches) {
-    if (min === 0) {
-      min = row.order;
-    } else {
-      min = Math.min(min, row.order);
-    }
-    max = Math.max(max, row.order);
-  }
-  return { min, max };
-});
-
-const isFirstRow = (order: number) => order === minMaxOrder.value.min;
-const isLastRow = (order: number) => order === minMaxOrder.value.max;
-
-const updateOrder = (researchIdx: number, researchPk: number, researchOrder: number, action: string) => {
-  if (action === 'inc_order' && researchOrder < minMaxOrder.value.max) {
-    const researchNearbyPk = props.tube.researches[researchIdx + 1].pk;
-    emit('updateOrder', { researchPk, researchNearbyPk, action });
-  } else if (action === 'dec_order' && researchOrder > minMaxOrder.value.min) {
-    const researchNearbyPk = props.tube.researches[researchIdx - 1].pk;
-    emit('updateOrder', { researchPk, researchNearbyPk, action });
-  } else {
-    root.$emit('msg', 'error', 'Ошибка');
-  }
-};
-
-const changeVisibility = (researchPk: number) => {
-  emit('changeVisibility', { researchPk });
-};
-
-const edit = (researchPk: number) => {
-  emit('edit', { researchPk });
-};
 </script>
 
 <style scoped lang="scss">
+.tube-item {
+  display: flex;
+}
 .sq {
     padding: 3px;
     display: inline-block;
@@ -144,87 +47,5 @@ const edit = (researchPk: number) => {
 .color-sq {
     height: 12px;
     width: 12px;
-}
-.tube-item {
-  display: flex;
-
-}
-.table {
-  table-layout: fixed;
-}
-.border {
-  border: 1px solid #bbb;;
-}
-.research {
-  background-color: #fff;
-  padding: 5px;
-  margin: 10px;
-  border-radius: 4px;
-  box-shadow: 0 1px 3px rgba(0, 0, 0, 0.12), 0 1px 2px rgba(0, 0, 0, 0.24);
-  transition: all 0.2s cubic-bezier(0.25, 0.8, 0.25, 1);
-  position: relative;
-
-  &.rhide {
-    background-image: linear-gradient(#6c7a89, #56616c);
-    color: #fff;
-  }
-
-  &:hover {
-    box-shadow: 0 14px 28px rgba(0, 0, 0, 0.25), 0 10px 10px rgba(0, 0, 0, 0.22);
-    z-index: 1;
-    transform: scale(1.008);
-  }
-}
-.research:not(:first-child) {
-  margin-top: 0;
-}
-
-.research:last-child {
-  margin-bottom: 0;
-}
-
-.research-title {
-  white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  padding-left: 3px;
-  padding-top: 1px;
-  padding-bottom: 1px;
-}
-
-.button {
-  width: 100%;
-  display: flex;
-  flex-wrap: nowrap;
-  flex-direction: row;
-  justify-content: stretch;
-}
-
-.transparent-button {
-  background-color: transparent;
-  align-self: stretch;
-  flex: 1;
-  color: #434A54;
-  border: none;
-  padding: 1px 0;
-
-}
-.transparent-button:hover {
-  background-color: #434a54;
-  color: #FFFFFF;
-  border: none;
-}
-.transparent-button:active {
-  background-color: #37BC9B;
-  color: #FFFFFF;
-}
-.transparent-button-disabled {
-  color: #abaeb3;
-  cursor: not-allowed;
-  background-color: transparent;
-  align-self: stretch;
-  flex: 1;
-  border: none;
-  padding: 1px 0;
 }
 </style>
